+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/view/qc/DetailedQcNode.java
@@ -77,9 +77,4 @@ public abstract class DetailedQcNode implements QcNode {
     return getDetailedQcStatusNote();
   }
 
-  @Override
-  public String getRunStatus() {
-    return null;
-  }
-
 }