+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/view/qc/PoolQcNode.java
@@ -103,11 +103,6 @@ public class PoolQcNode implements QcNode {
   }
 
   @Override
-  public String getRunStatus() {
-    return null;
-  }
-
-  @Override
   public List<RunQcNode> getChildren() {
     return getRuns();
   }