+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/view/qc/QcNode.java
@@ -25,8 +25,6 @@ public interface QcNode extends Serializable {
 
   public String getQcNote();
 
-  public String getRunStatus();
-
   public List<? extends QcNode> getChildren();
 
 }