+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/view/qc/RunPartitionAliquotQcNode.java
@@ -170,11 +170,6 @@ public class RunPartitionAliquotQcNode implements QcNode {
   }
 
   @Override
-  public String getRunStatus() {
-    return null;
-  }
-
-  @Override
   public List<? extends QcNode> getChildren() {
     return null;
   }